test: cover analytical leader producer floor
